refactor(hive-scout): split ingest-500-agents into cohesive modules
ingest-500-agents.ts was 1004 LOC, baselined at 1005 in the Module Size
Guard ratchet. Extract cohesive helpers so it drops to 446 LOC and leaves
the baseline entirely (61 files over 800 now, was 62):

- catalog-readme.ts — upstream README parsing (pure, no I/O)
- gap-mapping.ts — value-stream mapping, dedupe keys, gap detection,
  backlog-body rendering
- ambiguity-review.ts — bounded autonomous review: decision schema,
  metrics helpers, cache lookup, TAK-routed reviewer
- ingest-db.ts — narrowed Prisma surface, workforce links, admin notify

Pure code motion: no behavior change; the public API of
ingest-500-agents.ts is preserved via re-exports, so all importers
(tests, MCP tools, queue function, manual-run script) are untouched.
Hive-scout vitest suites pass (30/30).

Process-Spine-Decision: behavior-preserving module-size refactor, pure code motion with public API re-exported unchanged; no new capability, contract, or doc surface affected.
